requirements.txt
openai
pandas
scikit-learn
python-dotenv
numpy
tiktoken

